pyproject.toml
[tool.poetry]
name = "as3ninja"

version = "0.5.0"  # as3ninja/__init__.py -> __version__
license = "ISC"
description = "AS3 Ninja is a templating and validation engine for your AS3 declarations providing a CLI and Swagger REST API"
readme = "README.md"

authors = ["Simon Kowallik <github@simonkowallik.com>"]
maintainers = ["Simon Kowallik <github@simonkowallik.com>"]

homepage = "https://github.com/simonkowallik/as3ninja"
repository = "https://github.com/simonkowallik/as3ninja"
documentation = "https://as3ninja.readthedocs.io/"

keywords = [
    "F5",
    "AS3",
    "DevOps",
    "Security",
]

classifiers = [
    "License :: OSI Approved :: ISC License (ISCL)",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python",
    "Operating System :: OS Independent",
    "Natural Language :: English",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Topic :: Internet",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Version Control :: Git",
    "Topic :: Software Development :: Version Control",
    "Topic :: Software Development",
    "Topic :: System :: Installation/Setup",
    "Topic :: System :: Networking",
    "Topic :: System :: Systems Administration",
    "Topic :: System",
    "Topic :: Security",
]

include = [
    "docs/*",
    "docs/_static/*",
    "tests/*",
]

[tool.poetry.scripts]
as3ninja = "as3ninja.cli:cli"

[tool.poetry.dependencies]
python = "^3.6"
hvac = "^0.9.6"
jinja2 = "^2.10.3"
jsonschema = "^3.2.0"
loguru = ">=0.3.2, <1"
pydantic = "^1.2"
pyyaml = "^5.1"
six = "^1.14.0"
click = "^7.0"
fastapi = ">=0.48.0, <1"
uvicorn = ">=0.10.8, <1"

[tool.poetry.dev-dependencies]
pytest = "^5.3"
requests = "^2.20.0"
mock = "^4.0.0"
coverage = "^5.0.0"
pytest-cov = "^2.8.0"
pytest-mock = "^2.0.0"
Sphinx = "^2.4.0"
sphinx-autodoc-typehints = "^1.10.0"
sphinx_rtd_theme = "^0.4.0"
black = "^19.3b0"
isort = "^4.3"
mypy = "^0.761"
pylint = "^2.4"

[tool.poetry.urls]
issues = "https://github.com/simonkowallik/as3ninja/issues"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
